+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [CookieConsent] heading typography changes
 - [Accordion] heading typography changes
 - [Pagination] Active element changed from "a" element to "span" element
+- [Hero] content spacing changes
 
 #### Added